Reorganise datasets cache code
