+++ frontend/src/utils/cache/operation.ts
@@ -1,9 +1,15 @@
+import axios from 'axios';
 import * as localForage from 'localforage';
 import moment from 'moment';
 import { localForageKeys } from '..';
 import { fetchOperationById } from '../../pages/QueryData/utils';
 import { FetchOptions } from '../../types/api';
 import { OperationData } from '../../types/operations';
+import { OperationDataAPIResponse } from '../../types/operations';
+
+export interface QueryResult {
+  data: OperationDataAPIResponse;
+}
 
 export const getCachedOperationData = async ({
   limit,
@@ -50,4 +56,36 @@ export const isOperationDataCacheExpired = async ({
   }
 
   return true;
+};
+
+export const fetchOperations = async (path: string): Promise<QueryResult> => {
+  const token = await localForage.getItem<string>(localForageKeys.API_KEY);
+  const headers = {
+    'Content-Type': 'application/json',
+    Authorization: `token ${token}`,
+  };
+
+  return await axios(path, { headers });
+};
+
+export const isOperationsCacheExpired = async (
+  dataCountkey: string,
+  cacheDatasets: string,
+  path: string,
+): Promise<boolean> => {
+  const cachedDataCount: number | null = await localForage.getItem<number>(dataCountkey);
+  const cacheData = JSON.parse(cacheDatasets);
+  if (cachedDataCount) {
+    const { data } = await fetchOperations(path);
+    if (
+      data.count > cachedDataCount ||
+      JSON.stringify(cacheData.results) !== JSON.stringify(data.results)
+    ) {
+      localForage.setItem(dataCountkey, data.count);
+
+      return true;
+    }
+  }
+
+  return false;
 };
